backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt<4.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.13",
    "jwt>=1.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "pydotenv>=0.0.7",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
